requirements.txt
streamlit==1.32.0
pandas
numpy
scikit-learn
xgboost==1.7.6
shap==0.42.1
plotly
matplotlib
firebase-admin
PyPDF2
pdfplumber
PyGithub
nltk
google-generativeai
fpdf2
joblib
requests
imbalanced-learn
gdown
